requirements.txt
selenium==4.21.0
fastapi==0.111.0
pydantic-settings==2.2.1
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
lxml==5.2.2